// frontbox-pin-console/web/src/types/generated/Hardware.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DriverLookup } from "./DriverLookup";
import type { LedLookup } from "./LedLookup";
import type { ResolvedExpansionBoard } from "./ResolvedExpansionBoard";
import type { ResolvedIoBoard } from "./ResolvedIoBoard";
import type { SwitchLookup } from "./SwitchLookup";

export type Hardware = { switches: SwitchLookup, drivers: DriverLookup, leds: LedLookup, io_network: Array<ResolvedIoBoard>, exp_network: Array<ResolvedExpansionBoard>, };